Create p2.c
Write a C program to create a structure which will store the information about the
students (Roll No, Name, and Percentage of Marks).
a. Write a function to add a new record.
b. Write a function to display all existing record.
c. Write a function to search a record for any condition. d. Write a function to modify a record.