Create p5.c
Write a menu driven C program to create a singly connected linked list and do the following operations on it:
a. Write a function to insert a node at the beginning.
b. Write a function to insert a node at the end.
c. Write a function to insert a node at a specified position.
d. Write a function to delete a node from a specified position.
e. Write a function to count the number of nodes in the linked list.
f. Write a function to display the reverse linked list.